build: delete real, add fixed
Originally, I was going to pursue building with optionally either fixed
point or floating point, but between the hurdles necessary to support
both as a configuration option, the difference between the approximation
coefficients, and the pressure I expect from upstream for even
mentioning the FPU, it's not worth it. This is all fixed point.