build: simplify references to driver
The build now refers to the driver consistently as the driver, rather
than the module or kernel module. This was mostly a documentation
change, though some of the identifiers in the CMake script were updated
to reflect this.

Since there was sizable already a documentation change, I updated it
for the upcoming change to unify the libraries, which requires tracking
all of the kernel headers manually.