build: combine lib and common_lib
The original reason not to combine these was to keep the c lib separate
from the c++ lib, but just including those files from the c++ lib
already exposes all of the trouble of mixed compilation, so this
separation isn't useful.

Along the way, we started using the common lib's natural header
tracking as part of the trigger to kick off the Kbuild. Instead, we
track the headers manually. There's no technical duplication - each
header is only specified in the header list - but where driver headers
are specified in the CMake file is different from where the driver
source files are specified in the Kbuild. This is unavoidable short
of globbing.