feat(fixed): add curves_fixed_multiply
This revision adds fixed multiply.

Originally, the basis for fixed multiply was going to be
curves_mul_i64_i64_shr. Since adding that, though, I realized that we
shouldn't target Q31.32 as a sort of baseline format anywhere; all
operations should be arbitrary precision because the precision changes
at every step in these calculations. This means we need to be able
to also shift left while still at 128-bits, and that means
curves_mul_i64_i64_shr will be coming out.

This revision ports the old tests for curves_mul_i64_i64_shr and
complements it with a few new negative shift and boundary tests.