test(fixed): add new multiplication test cases
This revision adds test cases to exercise other parameterizations of
curves_fixed_multiply. The internal calculation ends up being the same
but how it arrives there is different.

These tests are analogous to testing multiple doors to the same room.
The result is the same, but you still have to test each of the doors.