build: add dependency from cmake onto kbuild files
I modified a kbuild file and cmake didn't kick it off. Turns out, it
only knows to invoke make on the dir. It wasn't checking the files
themselves. I've added them as explicit dependencies to the kbuild step.

Because the section they were under used to be called "Driver Headers",
and these aren't driver headers, that heading needed to be changed.
Because the section just before that was already called "Project Files",
I just removed "Driver Headers" and now these are all correctly labeled
as project files.