test(fixed): disable multiply and divide tests
Round to nearest isn't panning out the way I thought. The number of
special cases and the number of tests is too large. Plus, it's unlikely
to really help with accuracy since I'm already using an accumulator at
the end. Round to negative infinity is not the way to go, either, since
the errors are asymmetric. Instead, round to zero with an accumulator
is the middle ground. Its bias is the same for positive and negative
results, and it isn't nearly as computationally intensive as round to
nearest.

To get there, though, just about all of the tests for multiply and
divide are about to stop working in a nontrivial way. The fundamental
test plan has to change. Luckily, it will become easier because these
functions are going to be broken up into more fine-grained layers, but
these tests, as they stand, will be too difficult to make work.

This revision disables them in preparation. They're still here because
much of what's there will move, but they will just start failing in the
next few revisions unless they are disabled now.