refactor(fixed): improve parameter order
The order (frac_bits, value) made sense because the frac_bits are, in
one sense, part of the values type, the context needed to interpret
it. But every time it's used, it's (value << frac_bits), and when
talking about them, it's always the value that comes first.

This fits english language rules better, and frac_bits stops jumping
sides.