fix(fixed): fix overshoot in truncation
The previous version was only correct when shift == frac_bits. It would
overshoot in all other cases because it was biasing by just under one
at the original precision, instead of the target, reduced precison. It
was hard to tell because the values would still truncate correctly if
you converted them to integer, but the specific value didn't make sense
in tests, and the value after converting to double was always slightly
too large.

The fix is to bias by one at the target precision. This means frac_bits
isn't used, so it has been removed.