test(fixed): simplify tests after refactor
This test is originally from before the refactor, when rescale64 was
doing the error handling in line. Now that the error handling has moved
back to a cold function, this test does too much and blurs the line of
what it's testing. This new version only verifies that the error handler
is called, and that it is called with the correct parameters.