test(fixed): update from double tests
This is mostly a clean up pass to split the test vectors. Along the way,
I gave it a dedicated test param with values in the right order instead
of repurposing the to double param. I added an ostream inserter for
the param, streamlined the comments, and added a few new cases.