feat(fixed): rough in rne in rescale
This revision is a bit rough. It replaces the rtz in rescale with rne.
It fixes tests that broke, but with a very heavy hand. The spirit of
a few tests no longer match what they're actually testing. These will
need to be reviewed:
  add_rounding
  multiply_rounding
  rescale_s128_right_shift
  rescale_s64_right_shift
  subtract_rounding
  subtract_intermediate_then_downscale
  round_ints_negative
  round_ints_positive
  round_ints_edge_case_1
  round_ints_edge_case_32
  round_ints_edge_case_61

There are no tests specifically for rne. We can probably port many of
the rtz tests, but there are a few that are specific to rtz that don't
make much sense for rne. We also need tests specifically for rne
behaviors that aren't covered by the current rtz tests.

need tests for:
  __curves_fixed_shr_rne_s64
  __curves_fixed_shr_rne_s128