refactor(fixed): reorder conditional by cost
quotient == U64_MAX has no dependencies and is quick to evaluate. It
will tend to be false almost 100% of the time in production.

carry != 0 depends on the entire rest of the function and is roughly
50/50.

Evaluating quotient first is much cheaper and tends to short circuit
the rest of the evaluation.