refactor(fixed): improve isqrt initial guess
This revision uses a proper linear approximation of 1/sqrt(x) to find
the initial guess. Previously, we were only using 1 constant and a
hardcoded slope of -2 because we only needed a shift and a difference
to calculate it. Allowing the slope to vary requires an extra
multiplication, but it saves an entire NR iteration for the same
precision, which is 3 128-bit multiplications.