feat(spline): test geometric series and T(x)
There are valid parameterizations of the synchronous curve where the
first derivative explodes at 0. One is present here in the ui. In this
case, we literally cannot use cubic splines to approximate the function.
Instead, we are now approximating the transfer function, T(x) = xS(x),
which is well behaved and has valid derivatives.

We're investigating whether a linear set of buckets or a geometric
progression from small to large buckets with linear subdivisions
produces better error metrics.