test(spline): try gentle growth factor
This revision uses a fast log approximation to shift the size of
segments so they are smaller near 0.

The problem is, the fast log approximation ends up as part of both
the length of the tangents and causes us to sometimes choose the wrong
segment. Overall, this is much less precise.