feat(spline): restore scale to the curves
I've been debating whether to put scale in the curves or make it a
generic parameter. If we put it in the curves, we have to manage it in
each curve, but some curves may actually need to know this. Putting it
in the curves means we don't have to mess with separately it in the
approximation.

For now, put it in the curves.