refactor(config): switch to concrete structs
Previously, we were using bags of params to represent curves. This is
fine from a purely data-driven perspective, but there were too many
places we had to step in and interpret the data concretely,
particularly when mapping between enums for curve data and the curves
themselves. This revision still uses parameters at the leaves, but
switches to concrete structs for the individual curves and their
parameters.