refactor(spline): add dedicated value path
When integrating, we can't use the derivative near 0 when the
parameterization gives and asymptote, so we either need to use gaussian
everywhere, which doesn't need a derivative, or use a hybrid approach
where we start with gaussian then switch to corrected trapezoid after
the asymptote dies down.

There's no real, clear switching critera, so we'll just use gaussian
everywhere on the curves that may need it. gauss5 is actually higher
accuracy than even 2 richardsons, but richardsons gets to reuse the
tangents from the knots for free. In our use, the difference is likely a
blip.