feat(spline): bleed off much more gently
Previously, we bled off over the domain of just one segment one past the
final octave. It was too fast. Now we bleed off over an entire octave.
It takes some finagling if the curve hasn't flattend out by the time
it reaches the geometric limit - we may have to pull it out of a dive -
but overall, this works well as the final safety mechanism. We'll have
an optional soft limiter for curves that don't straighten out.