fix(spline): enforce c0 continuity
There were a few jumps near 0 in sensitivity. They are now gone.

This revision forces d to be the sum of the previous abcd, which is the
literal invariant of a spline. This affects error a little, but the
visual effect is strong. Lots of little misalignments now gone and even
the noise near x=0 seems to line up better.